requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-simple-history==3.10.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.3